[pytorch][ci] add custom selective build flow for android build
Summary:
Mobile custom selective build has already been covered by `test/mobile/custom_build/build.sh`.
It builds a CLI binary with host-toolchain and runs on host machine to
check correctness of the result.

But that custom build test doesn't cover the android/gradle build part.
And we cannot use it to measure and track the in-APK size of custom
build library.

So this PR adds the selective build test coverage for android NDK build.
Also integrate with the CI to upload the custom build size to scuba.

ghstack-source-id: 20cc40487d018c948740546716c9a23f761aa6c9
Pull Request resolved: https://github.com/pytorch/pytorch/pull/40199